竞赛证书识别系统/requirements.txt
streamlit>=1.28.0
sqlmodel>=0.0.14
bcrypt>=4.0.1
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
requests>=2.31.0
python-dotenv>=1.0.0
fpdf2>=2.7.0

